Fixed bug #25316 (Possible infinite loop inside _php_stream_write()).

+++ main/streams/streams.c
@@ -867,7 +867,8 @@ static size_t _php_stream_write_buffer(php_stream *stream, const char *buf, size
 
 		justwrote = stream->ops->write(stream, buf, towrite TSRMLS_CC);
 
-		if (justwrote > 0) {
+		/* convert justwrote to an integer, since normally it is unsigned */
+		if ((int)justwrote > 0) {
 			buf += justwrote;
 			count -= justwrote;
 			didwrite += justwrote;
